+++ src/App.js
@@ -1,9 +1,9 @@
 import { useState } from 'react';
 import { Routes, Route, useNavigate, Navigate } from 'react-router-dom';
 import * as authService from './utility/auth';
-import Landing from './pages/Landing/Landing';
-import Signup from './pages/Signup/Signup';
-import Login from './pages/Login/Login';
+import Landing from './pages/Landing';
+import Signup from './pages/Signup';
+import Login from './pages/Login';
 import Home from './pages/Home/Home';
 import Onboarding from './pages/Onboarding/Onboarding';
 import Expenses from './pages/Expenses/Expenses';
@@ -25,7 +25,7 @@ function App() {
   }
 
   return (
-    <div className="App">
+    <div>
       <Routes>
         {user && localStorage.getItem("token") ? (
           <>
