--- src/components/LandingNavbar/LandingNavbar.module.css
@@ -1,26 +1,0 @@
-.navbar {
-    display: flex;
-    justify-content: space-between;
-    background-color: coral;
-}
-
-.logo {
-    padding: 10px;
-    align-self: center;
-}
-
-.buttons {
-    padding: 10px;
-    justify-content: center;
-}
-
-.login {
-    background-color: coral;
-    border: none;
-    padding: 10px;
-}
-
-.signup {
-    padding: 10px;
-    border-radius: 0.5em;
-}
