--- src/components/LoginForm/LoginForm.module.css
@@ -1,53 +1,0 @@
-.container {
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 50px;
-}
-
-input {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    position: static;
-    width: 248px;
-    height: 39px;
-    left: 0px;
-    top: 23px;
-    background: #bcbcbc;
-    border-radius: 5px;
-}
-
-.signinButton {
-    margin-top: 18px;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    position: static;
-    width: 248px;
-    height: 39px;
-    left: 0px;
-    background: #737373;
-    border-radius: 5px;
-}
-
-.label {
-    position: static;
-    width: 131px;
-    height: 19px;
-    left: 0px;
-    top: 0px;
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 16px;
-    line-height: 18.75px;
-}
-
-
-.inputContainer {
-    margin-top: 18px;
-}
