--- src/components/SignupForm/SignupForm.module.css
@@ -1,54 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
-
-.container {
-    display: flex;
-    flex-flow: column nowrap;
-    justify-content: center;
-    align-items: center;
-    margin-bottom: 50px;
-}
-
-.signupButton {
-    margin-top: 18px;
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    position: static;
-    width: 248px;
-    height: 39px;
-    left: 0px;
-    background: #737373;
-    border-radius: 5px;
-}
-
-.label {
-    position: static;
-    width: 131px;
-    height: 19px;
-    left: 0px;
-    top: 0px;
-    font-family: Roboto;
-    font-style: normal;
-    font-weight: normal;
-    font-size: 16px;
-    line-height: 18.75px;
-}
-
-input {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    position: static;
-    width: 248px;
-    height: 39px;
-    left: 0px;
-    top: 23px;
-    background: #bcbcbc;
-    border-radius: 5px;
-}
-
-.inputContainer {
-    margin-top: 18px;
-}
