--- src/pages/Landing/Landing.module.css
@@ -1,52 +1,0 @@
-/* .container {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-} */
-
-.landingIntro {
-    display: flex;
-    flex-direction: column;
-    height: 200px;
-    align-items: center;
-    background-color: aqua;
-    width: 100%;
-}
-
-.signupButton {
-    width: 200px;
-    padding: 0.5em;
-    border-radius: 0.5em;
-}
-
-.landingPageInfo {
-    width: 80%;
-    margin: 0 auto;
-}
-
-.featureInfo {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-top: 20px;
-}
-
-.featureImg {
-    display: block;
-    width: 150px;
-}
-
-.featureTitle {
-    margin: 0.25em 0;
-}
-
-.featureDescription {
-    text-align: center;
-    margin: 0.25em 0;
-}
-
-.featureButton {
-    margin: 50px 0 20px 0;
-    display: flex;
-    justify-content: center;
-}
