+++ src/pages/Signup.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import SignupForm from '../../components/SignupForm/SignupForm';
-import LandingNavbar from '../../components/LandingNavbar/LandingNavbar';
-import signup from '../../images/signup.svg';
+import SignupForm from '../components/SignupForm';
+import LandingNavbar from '../components/LandingNavbar';
+import welcome from '../images/welcome.svg';
 
 const Signup = (props) => {
   const [message, setMessage] = useState([""]);
@@ -15,7 +15,7 @@ const Signup = (props) => {
       <LandingNavbar/>
       <main class="column">
         <h1 className='logo'>Welcome to ea<span>$</span>e</h1>
-        <img src={signup} alt="ease logo"/>
+        <img src={welcome} alt="ease logo"/>
         <p>{message}</p>
         <SignupForm {...props} updateMessage={updateMessage} />
       </main>
