--- src/pages/Signup/Signup.module.css
@@ -1,14 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
-
-.signup {
-    display: flex;
-    align-items: center;
-    flex-direction: column;
-}
-
-img {
-    align-self: center;
-    margin-top: 18px;
-    width: 248px;
-    height: 144px;
-}
